Moved from common to local spawn vm getting rid of common example file.
