+++ runtime/src/lib.rs
@@ -7,3 +7,4 @@ pub mod local_notification_handler;
 mod response_parser;
 pub mod vm;
 pub mod vm_runner;
+pub mod local_spawn_vm;
